src/collector/requirements.txt
torch==2.9.0
torchvision==0.24.0
pillow==11.3.0
uvicorn==0.38.0
requests==2.32.5
fastapi==0.121.1
tqdm==4.67.1